service e controller de categoria feitos

+++ manutencao-spring/manutencao/src/main/java/br/net/dwii/projeto/manutencao/model/Orcamento.java
@@ -3,14 +3,14 @@ package br.net.dwii.projeto.manutencao.model;
 import java.util.List;
 
 public class Orcamento {
-    public Long id;
-    public long idSolicitacao;
+    public int id;
+    public int idSolicitacao;
     public List<Servico> servico;
     public Double valorTotal;
     public boolean aprovado;
     public String msgRejeicao;
     
-    public Orcamento(long id, long idSolicitacao, List<Servico> servico, Double valorTotal, boolean aprovado,
+    public Orcamento(int id, int idSolicitacao, List<Servico> servico, Double valorTotal, boolean aprovado,
             String msgRejeicao) {
         this.id = id;
         this.idSolicitacao = idSolicitacao;
@@ -20,19 +20,19 @@ public class Orcamento {
         this.msgRejeicao = msgRejeicao;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public long getIdSolicitacao() {
+    public int getIdSolicitacao() {
         return idSolicitacao;
     }
 
-    public void setIdSolicitacao(long idSolicitacao) {
+    public void setIdSolicitacao(int idSolicitacao) {
         this.idSolicitacao = idSolicitacao;
     }
